+++ R/create_assoc_args.R
@@ -17,7 +17,8 @@ create_assoc_args <- function(
         "--assoc",
         "--maf", assoc_params$maf,
         "--ci", assoc_params$confidence_interval,
-        "--out", assoc_params$base_output_filename
+        "--out", assoc_params$base_output_filename,
+        "--noweb" # Will freeze otherwise
       )
     )
   }