+++ R/create_plink_v2_0_options.R
@@ -21,7 +21,6 @@ create_plink_v2_0_options <- function(
   plinkr::create_plink_options(
     plink_exe_path = plink_exe_path,
     plink_version = plink_version,
-    os = os,
-    add_noweb = FALSE # Option does not exist in PLINK2
+    os = os
   )
 }