+++ tests/testthat/test-read_plink_phenotype_file.R
@@ -55,3 +55,21 @@ test_that("multiple traits", {
   expect_equal(6, nrow(phenotypes))
   file.remove(phenotype_filename)
 })
+
+test_that("read PLINK tutorial file pop.phe", {
+  expect_equal(1 + 1, 2) # Prevents testthat warning for empty test
+  if (!is_plink_tutorial_data_installed()) return()
+  phenotype_filename <- stringr::str_subset(
+    get_plink_tutorial_data_filenames(), "pop.phe"
+  )
+  expect_silent(read_plink_phenotype_file(phenotype_filename))
+})
+
+test_that("read PLINK tutorial file qt.phe", {
+  expect_equal(1 + 1, 2) # Prevents testthat warning for empty test
+  if (!is_plink_tutorial_data_installed()) return()
+  phenotype_filename <- stringr::str_subset(
+    get_plink_tutorial_data_filenames(), "qt.phe"
+  )
+  expect_silent(read_plink_phenotype_file(phenotype_filename))
+})